+++ NEWS.md
@@ -1,3 +1,5 @@
+# santoku (development version)
+
 # santoku 0.8.0
 
 ## Breaking changes